MOB-458 (#177)
Improved coroutines and scopes.
Added suspend to all functions in datasources, and ensured they are main-safe.
Ensured all read and write from SharedPreferences is done in a different thread.
Added a dispatcher in repository getEventAttendees as it may be computationally expensive because it processes a list of components and therefore moving to a worked thread can be beneficial.
Configured StrictMode.