MOB-465 (#179)
Refactored EventDetails: 
Started by removing parameters not needed in UiState. Some of them meant having login inside the state and it should be avoided. Also renamed some parameters to better express what they represent.
Deleted EventDetailsUiStateHelper, as everything that may be needed should be done in the mapper.
Removed some variables in ViewModel and added an internal event to keep the changes that user may make.